Add files via upload
Just created the forwards propagation of a CNN. Spent most of my time reading up on them and I think I get the point. Shouldn't take too long to create the backpropagation and then move on to the evolution engine that takes a generation of nets with fitnesses and creates the next generation out of them.